% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.support-it.R
\name{lang.support.it}
\alias{lang.support.it}
\title{Language support for Italian}
\usage{
lang.support.it(...)
}
\arguments{
\item{...}{Optional arguments for \code{\link[koRpus:set.lang.support]{set.lang.support}}.}
}
\description{
This function adds support for Italian to the koRpus package. You should not
need to call it manually, as that is done automatically when this package is
loaded.
}
\details{
In particular, this function adds the following:
\itemize{
 \item \code{lang}: The additional language "it" to be used with koRpus
 \item \code{treetag}: The additional preset "it",
      implemented according to the respective TreeTagger[1] script
 \item \code{POS tags}: An additional set of tags,
      implemented using the documentation for the corresponding
   TreeTagger parameter sets[2, 3]
}

To also be able to use hyphenation, the package sylly.it needs to exist and be loaded.
}
\examples{
\dontrun{
lang.support.it()
}
}
\references{
[1] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/}

[2] \url{http://www.cis.uni-muenchen.de/~schmid/tools/TreeTagger/data/italian-tagset.txt}

[3] \url{http://sslmit.unibo.it/~baroni/collocazioni/itwac.tagset.txt} (alt. tagset)
}
